+++ src/commands/look/responses.js
@@ -1,10 +1,14 @@
 const { aResponse } = require('../../scure-response');
 const { getPossibleDestinationsSentence } = require('../../scure-commons');
 const { getDescription } = require('../../scure-commons');
+const { handlePluginExtensions } = require('../../lib/plugin-executor');
 
 const itemNotFound = scure => aResponse(scure.sentences.get('item-not-in-location'));
-const itemDescription = (scure, data, item) =>
-  aResponse(getDescription(item.description, data, scure));
+const itemDescription = (scure, data, item) => {
+  const description = getDescription(item.description, data, scure);
+  return aResponse(handlePluginExtensions(description, data, scure));
+};
+
 const lookRoom = (scure, data, roomId) => {
   const room = scure.rooms.getRoom(roomId);
   const description = getDescription(room.description, data, scure);